+++ sync-basalam/includes/Plugin.php
@@ -15,7 +15,7 @@ defined('ABSPATH') || exit;
 
 class Plugin
 {
-    public const VERSION = '1.8.8';
+    public const VERSION = '1.9.0';
 
     public function __construct()
     {